+++ src/app/shared/service/api.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { ApiService } from './api.service';
+
+describe('ApiService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [ApiService]
+    });
+  });
+
+  it('should be created', inject([ApiService], (service: ApiService) => {
+    expect(service).toBeTruthy();
+  }));
+});
